Move sessionToBinary and beginOfWeek to utils
